Silence spurious clang implicit 64 to 32 bit truncation warnings in trunnel
clang warns that a few implicit casts in trunnel truncate values from
64 to 32 bit. We do explicit casts to silence these warnings.

+++ src/ext/trunnel/trunnel.c
@@ -75,7 +75,8 @@ uint64_t
 trunnel_htonll(uint64_t a)
 {
 #ifdef IS_LITTLE_ENDIAN
-  return trunnel_htonl(a>>32) | (((uint64_t)trunnel_htonl(a))<<32);
+  return trunnel_htonl((uint32_t)(a>>32))
+    | (((uint64_t)trunnel_htonl((uint32_t)a))<<32);
 #else
   return a;
 #endif
